Triad.py is a python program that utilizes a fundamental theorem used in math, computer science, physics and engineering called the parallelogram law.
The resulting software program produces phasors which are rotating vectors.  Now the program has two vectors rotating relative to each other and the resultant vector has its own angular velocity.  Currently, the software utilizes turtle graphics that statically graphs the position as well as the length of each vector. The program itself calculates the individual and relative velocities.  The equations for the program were derived with the length and rotational velocities as variables and are input into the start of the program so that the researcher can examine a particular path of rotation producing position and velocity measurements.
   Future modifications of the program include but are not limited to calculations for accelerations of the resultant vector and weighted members for the vectors themselves giving the program a connection to real life systems, not merely theoretical.  One interesting aspect of observing rotating vectors is that even though each vector velocity rotates in the chosen positive direction, the relative velocity of the resultant vector can be observed to rotate in the negative direction, which creates an unexpected and counter intuitive experience for the observer.
    One current theoretical application of this software is in the area of computer security and secure messaging as it applies to one time pad messages.  Each time a new message can be sent after changing the initial parameters of vector length and velocity causing the message to be viewed as a one time pad message, so there are no duplicate messages, using the same computer code.